requirements.txt
pandas==2.2.3
pyhive==0.7.0
pyyaml==6.0.2

